tests/krate/publish: Extract `readme` module
